requirements.txt
tensorflow[and-cuda]
keras
pandas
numpy
scikit-learn
pyzmq
opencv-python
Pillow